TFP-3693: Refaktorert finn andel av beregnet.
