+++ src/main/java/no/nav/folketrygdloven/beregningsgrunnlag/arbeidstaker/BeregnPrArbeidsforholdFraAOrdningen.java
@@ -66,42 +66,7 @@ class BeregnPrArbeidsforholdFraAOrdningen extends LeafSpecification<Beregningsgr
      * @return Andel av snitt fra a-ordningen
      */
     private BigDecimal finnAndelAvBeregnet(BigDecimal beregnetPrÅr, BeregningsgrunnlagPrArbeidsforhold arbeidsforhold, BeregningsgrunnlagPeriode periode) {
-        YtelsesSpesifiktGrunnlag ytelsesSpesifiktGrunnlag = periode.getBeregningsgrunnlag().getYtelsesSpesifiktGrunnlag();
-        if (!arbeidsforhold.erFrilanser() && arbeidsforhold.getArbeidsgiverId() != null && ytelsesSpesifiktGrunnlag instanceof OmsorgspengerGrunnlag) {
-            return fordelRestinntektFraAOrdningen(beregnetPrÅr, arbeidsforhold, periode);
-        }
-        return beregnetPrÅr;
-    }
-
-    private BigDecimal fordelRestinntektFraAOrdningen(BigDecimal beregnetPrÅr, BeregningsgrunnlagPrArbeidsforhold arbeidsforhold, BeregningsgrunnlagPeriode periode) {
-        Inntektsgrunnlag inntektsgrunnlag = periode.getInntektsgrunnlag();
-        List<BeregningsgrunnlagPrArbeidsforhold> arbeidsforholdISammeOrg = periode.getBeregningsgrunnlagPrStatus(AktivitetStatus.ATFL)
-            .getArbeidsforholdIkkeFrilans()
-            .stream()
-            .filter(a -> a.getArbeidsgiverId() != null)
-            .filter(a -> a.getArbeidsgiverId().equals(arbeidsforhold.getArbeidsgiverId()))
-            .collect(Collectors.toList());
-        BigDecimal beløpFraInntektsmeldingPrÅr = arbeidsforholdISammeOrg.stream()
-            .map(inntektsgrunnlag::getInntektFraInntektsmelding)
-            .map(beløp -> beløp.multiply(BigDecimal.valueOf(12)))
-            .reduce(BigDecimal::add)
-            .orElse(BigDecimal.ZERO);
-
-        if (beløpFraInntektsmeldingPrÅr.compareTo(beregnetPrÅr) > 0) {
-            return BigDecimal.ZERO;
-        }
-
-        BigDecimal restFraAOrdningen = beregnetPrÅr.subtract(beløpFraInntektsmeldingPrÅr);
-
-        long antallArbeidsforholdUtenInntektsmelding = arbeidsforholdISammeOrg.stream()
-            .filter(a -> inntektsgrunnlag.getInntektFraInntektsmelding(a) == null || inntektsgrunnlag.getInntektFraInntektsmelding(a).compareTo(BigDecimal.ZERO) == 0)
-            .count();
-
-        if (antallArbeidsforholdUtenInntektsmelding == 0) {
-            throw new IllegalStateException("Kan ikke beregne andel fra aordningen når alle andeler for arbeidsgiver med orgnr " + arbeidsforhold.getArbeidsgiverId() +
-                " har inntektsmelding");
-        }
-
-        return restFraAOrdningen.divide(BigDecimal.valueOf(antallArbeidsforholdUtenInntektsmelding), RoundingMode.HALF_EVEN);
+        YtelsesSpesifiktGrunnlag getYtelsesSpesifiktGrunnlag = Objects.requireNonNull(periode.getBeregningsgrunnlag().getYtelsesSpesifiktGrunnlag(), "getYtelsesSpesifiktGrunnlag");
+        return getYtelsesSpesifiktGrunnlag.finnAndelAvBeregnet(beregnetPrÅr, arbeidsforhold, periode);
     }
 }
