+++ src/main/java/no/nav/folketrygdloven/beregningsgrunnlag/regelmodell/ytelse/YtelsesSpesifiktGrunnlag.java
@@ -1,8 +1,12 @@
 package no.nav.folketrygdloven.beregningsgrunnlag.regelmodell.ytelse;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonBackReference;
 
 import no.nav.folketrygdloven.beregningsgrunnlag.regelmodell.resultat.Beregningsgrunnlag;
+import no.nav.folketrygdloven.beregningsgrunnlag.regelmodell.resultat.BeregningsgrunnlagPeriode;
+import no.nav.folketrygdloven.beregningsgrunnlag.regelmodell.resultat.BeregningsgrunnlagPrArbeidsforhold;
 
 public abstract class YtelsesSpesifiktGrunnlag {
 
@@ -22,4 +26,16 @@ public abstract class YtelsesSpesifiktGrunnlag {
         this.beregningsgrunnlag = beregningsgrunnlag;
     }
 
+    /**
+     * I tilfelle der det er omsorgspenger og det er kun mottatt inntektsmelding for enkelt arbeidsforhold ho en arbeidsgiver
+     * skal resterende arbeidsforhold fordele restbeløpet fra a-ordningen mellom seg (https://jira.adeo.no/browse/TSF-1153).
+     *
+     * I alle andre caser gis det fulle snittbeløpet til dette arbeidsforholdet.
+     *
+     * @param beregnetPrÅr Snitt de siste 3 månedene fra a-ordningen
+     * @param arbeidsforhold Arbeidsforhold
+     * @param periode Periode
+     * @return Andel av snitt fra a-ordningen
+     */
+    public abstract BigDecimal finnAndelAvBeregnet(BigDecimal beregnetPrÅr, BeregningsgrunnlagPrArbeidsforhold arbeidsforhold, BeregningsgrunnlagPeriode periode);
 }
