+++ src/main/java/no/nav/folketrygdloven/beregningsgrunnlag/regelmodell/ytelse/fp/ForeldrepengerGrunnlag.java
@@ -1,5 +1,9 @@
 package no.nav.folketrygdloven.beregningsgrunnlag.regelmodell.ytelse.fp;
 
+import java.math.BigDecimal;
+
+import no.nav.folketrygdloven.beregningsgrunnlag.regelmodell.resultat.BeregningsgrunnlagPeriode;
+import no.nav.folketrygdloven.beregningsgrunnlag.regelmodell.resultat.BeregningsgrunnlagPrArbeidsforhold;
 import no.nav.folketrygdloven.beregningsgrunnlag.regelmodell.ytelse.YtelsesSpesifiktGrunnlag;
 
 public class ForeldrepengerGrunnlag extends YtelsesSpesifiktGrunnlag {
@@ -12,5 +16,10 @@ public class ForeldrepengerGrunnlag extends YtelsesSpesifiktGrunnlag {
 
     public boolean erBesteberegnet() {
         return erBesteberegnet;
+    }
+
+    @Override
+    public BigDecimal finnAndelAvBeregnet(BigDecimal beregnetPrÅr, BeregningsgrunnlagPrArbeidsforhold arbeidsforhold, BeregningsgrunnlagPeriode periode) {
+        return beregnetPrÅr;
     }
 }
