+++ src/main/java/no/nav/folketrygdloven/beregningsgrunnlag/regelmodell/ytelse/sp/SykepengerGrunnlag.java
@@ -1,5 +1,9 @@
 package no.nav.folketrygdloven.beregningsgrunnlag.regelmodell.ytelse.sp;
 
+import java.math.BigDecimal;
+
+import no.nav.folketrygdloven.beregningsgrunnlag.regelmodell.resultat.BeregningsgrunnlagPeriode;
+import no.nav.folketrygdloven.beregningsgrunnlag.regelmodell.resultat.BeregningsgrunnlagPrArbeidsforhold;
 import no.nav.folketrygdloven.beregningsgrunnlag.regelmodell.ytelse.YtelsesSpesifiktGrunnlag;
 
 public class SykepengerGrunnlag extends YtelsesSpesifiktGrunnlag {
@@ -8,4 +12,8 @@ public class SykepengerGrunnlag extends YtelsesSpesifiktGrunnlag {
         super("SP");
     }
 
+    @Override
+    public BigDecimal finnAndelAvBeregnet(BigDecimal beregnetPrÅr, BeregningsgrunnlagPrArbeidsforhold arbeidsforhold, BeregningsgrunnlagPeriode periode) {
+        return beregnetPrÅr;
+    }
 }
